feat(web_console): datasource and trainning
