+++ web_console/components/JobCommonInfo.jsx
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react';
 import css from 'styled-jsx/css';
-import { Table, Link, Text, Card, Description, Popover, useTheme } from '@zeit-ui/react';
+import { Table, Link, Text, Card, Description, Popover, useTheme, Button } from '@zeit-ui/react';
 import useSWR from 'swr';
 
 import { fetcher } from '../libs/http';
